Tracking node is working... kind of.
We can get the camera and presumably find the center of a ball, but I haven't called the functions
that display the color sliders yet. I've yet to implement the angle-finding thingie into our classes.

Here's how it'll all work: we'll start phi-keyboard-pilot AND phi-locate-circle in separate nodes.
phi-locate-circle will subscribe to a topic named "ardrone/autopilot" or something like that, a topic
which will be published to by phi-keyboard-pilot by the press of some key. If phi-locate-circle gets
a message in that topic, it'll start following the ball. When we press any key in phi-keyboard-pilot,
the "shouldFollow" variable in the locate circle class will be set to false, and control will be given
back to the user.

I'm, uh, using English because vim doesn't support Portuguese accent marks, and I hate writing in Por-
tuguese without using accent marks. Besides, it's cooler this way. HAHAH :D